Prototype: highlight editor updates from tobias
When I press “play selected clips” the scrubber is missing on the clips (showing me what is playing in the big window)
Free users should be able listen to all music but not select. Add the same premium icons here.
Make the track box clickable to play/pause the track.
Missing pause button on the big player when playing clips